+++ NEWS.md
@@ -3,9 +3,10 @@ raveio 0.0.4
 
 Major changes: 
 
-* removed `lazyarray`
+* removed `lazyarray`, `pryr`
 * fixed `get_ram` errors when system command not found
 * disabled support on `Solaris`
+* changed `hdf5r` to `rhdf5`
 
 raveio 0.0.3
 =======